+++ interfac/mysystem.h
@@ -14,7 +14,7 @@ extern "C" {
 
 class mystat;
 
-char *myfgets(char *, size_t, FILE *);
+char *myfgets(char *, int, FILE *);
 int mysystem(const char *);
 int mysystem2(const char *, const char *);
 char *mytmpdir(const char *);
